Handle better updates

+++ src/modules/aws_security_group@0.0.1/migration/1636587967230-aws_security_group.ts
@@ -7,6 +7,7 @@ export class awsSecurityGroup1636587967230 implements MigrationInterface {
         await queryRunner.query(`CREATE TABLE "aws_security_group" ("id" SERIAL NOT NULL, "description" character varying, "group_name" character varying, "owner_id" character varying, "group_id" character varying, "vpc_id" character varying, CONSTRAINT "PK_c95f0d51761dbddd9c3950d75be" PRIMARY KEY ("id"))`);
         await queryRunner.query(`CREATE TABLE "aws_security_group_rule" ("id" SERIAL NOT NULL, "security_group_rule_id" character varying, "is_egress" boolean NOT NULL, "ip_protocol" character varying NOT NULL, "from_port" integer, "to_port" integer, "cidr_ipv4" cidr, "cidr_ipv6" cidr, "prefix_list_id" character varying, "description" character varying, "security_group_id" integer, CONSTRAINT "PK_e9776ba4916babd78e029b421e4" PRIMARY KEY ("id"))`);
         await queryRunner.query(`ALTER TABLE "aws_security_group_rule" ADD CONSTRAINT "FK_6d3482619216803d2f14ecf609d" FOREIGN KEY ("security_group_id") REFERENCES "aws_security_group"("id") ON DELETE NO ACTION ON UPDATE NO ACTION`);
+        await queryRunner.query(`ALTER TABLE "aws_security_group_rule" ADD CONSTRAINT "UQ_rule" UNIQUE ("is_egress", "ip_protocol", "from_port", "to_port", "cidr_ipv4", "security_group_id")`);
         // Example of use: call create_or_update_aws_security_group('test', 'test', '[{"isEgress": false, "ipProtocol": "tcp", "fromPort": "8088", "toPort": 8088, "cidrIpv4": "0.0.0.0/0"}]', 'vpc');
         await queryRunner.query(`
             create or replace procedure create_or_update_aws_security_group(
@@ -20,6 +21,7 @@ export class awsSecurityGroup1636587967230 implements MigrationInterface {
                 declare
                     sec_group_id integer;
                     json_rule jsonb;
+                    rule_ids integer[];
                 begin
                     if _vpc_id is null then
                         select 'default' into _vpc_id;
@@ -49,9 +51,25 @@ export class awsSecurityGroup1636587967230 implements MigrationInterface {
                             description = _description
                         where id = sec_group_id;
 
-                        -- TODO: Handle better rules update
-                        delete from aws_security_group_rule
-                        where security_group_id = sec_group_id;
+                        if jsonb_array_length(_security_group_rules) > 0 then
+                            for json_rule in
+                                select * from jsonb_array_elements(_security_group_rules)
+                            loop
+                                assert json_rule ?& array['isEgress', 'ipProtocol', 'fromPort', 'toPort', 'cidrIpv4'], 'Not all security group rule required keys are defined ("isEgress", "ipProtocol", "fromPort", "toPort", "cidrIpv4")';
+
+                                perform array_append(rule_ids, id)
+                                from aws_security_group_rule
+                                where security_group_id = sec_group_id and
+                                    is_egress = cast(json_rule ->> 'isEgress' as boolean) and 
+                                    ip_protocol = json_rule ->> 'ipProtocol' and
+                                    from_port = cast(json_rule ->> 'fromPort' as integer) and
+                                    to_port = cast(json_rule ->> 'toPort' as integer) and
+                                    cidr_ipv4 = cast(json_rule ->> 'cidrIpv4' as cidr);
+                            end loop;
+
+                            delete from aws_security_group_rule
+                            where security_group_id = sec_group_id and not(id = any(rule_ids));
+                        end if;
                     end if;
 
                     if jsonb_array_length(_security_group_rules) > 0 then
@@ -63,7 +81,8 @@ export class awsSecurityGroup1636587967230 implements MigrationInterface {
                             insert into aws_security_group_rule
                                 (security_group_id, is_egress, ip_protocol, from_port, to_port, cidr_ipv4)
                             values
-                                (sec_group_id, cast(json_rule ->> 'isEgress' as boolean), json_rule ->> 'ipProtocol', cast(json_rule ->> 'fromPort' as integer), cast(json_rule ->> 'toPort' as integer), cast(json_rule ->> 'cidrIpv4' as cidr));
+                                (sec_group_id, cast(json_rule ->> 'isEgress' as boolean), json_rule ->> 'ipProtocol', cast(json_rule ->> 'fromPort' as integer), cast(json_rule ->> 'toPort' as integer), cast(json_rule ->> 'cidrIpv4' as cidr))
+                            on conflict do nothing;
                         end loop;
                     end if;
